Re-enable clang-tidy for core workflows + clang-tidy fixes (#197)
* Ensure the clang-tidy is updated + clang-tidy fixes

* update-ci workflow

* Enable clang-tidy checks

* Add extra logging to device counter collection samples

* Misc clang-tidy fixes

* Disable device counter collection samples for ThreadSanitizer

* Formatting

---------

+++ source/lib/rocprofiler-sdk-att/wave.cpp
@@ -62,10 +62,10 @@ WaveFile::WaveFile(WaveConfig& config, const att_wave_data_t& wave)
     for(size_t i = 0; i < wave.instructions_size; i++)
     {
         auto& inst = wave.instructions_array[i];
-        instructions.push_back({(int64_t) inst.time,
-                                (int) inst.category,
-                                (int) inst.stall,
-                                (int64_t) inst.duration,
+        instructions.push_back({inst.time,
+                                static_cast<int>(inst.category),
+                                static_cast<int>(inst.stall),
+                                static_cast<int64_t>(inst.duration),
                                 config.code->line_numbers[inst.pc]});
     }
